+++ src/main/kotlin/com/github/erjotek/stickyprojectfolder/startup/MyProjectActivity.kt
@@ -11,6 +11,5 @@ class MyProjectActivity : ProjectActivity {
         thisLogger().info("Installing StickyProjectFolder manager")
         val manager = StickyScrollManager(project)
         com.intellij.openapi.util.Disposer.register(project, manager)
-        manager.install()
     }
 }